cupy: complex: make real and imag to a property like in numpy
